Avoid incremental transform shifting during graph bar rendering
Instead of letting addition errors and cognitive load accumulate
simply use the same offset calculation that is used for lines and
reset the transform back to the top-left corner after every drawn bar.

It's conceptually simpler, and will be useful later for
stacked graphs and more.

It also fixes a discrepancy between plain and step_shape rendering.
Prior to this commit, bar coloring occurred with a disturbed transform,
which was centered at the last drawn bar, wherever it happened to be.
This meant that nontrivial colors, like gradients and images,
were arbitrarily offset when step_shape was used.