Draw shaped graph bars only up to the baseline
Prior to this commit, every step_shape bar was,
for some reason, drawn with height = full widget's height,
regardless of the value it represents.

So for the most bars their lower end was far outside the box
and therefore clipped off. So e.g. one couldn't see the
lower round edges of rounded rectangles.

This commit fixes that by appropriately varying the height
that is passed into step_shape(), so that lower edges of the bars
always exactly touch the lower edge of the graph.

If the user wants to restore the prior behavior, they can simply
ignore the passed-in height in step_shape() and
just always draw their shape with height = full widget height.

But they couldn't replicate the effect of this commit in
their rc.lua without essentially writing their own graph widget.