Fix border_color/border_width bugs in graph widget
add_value() wrongly assumed that border_width is always set,
when border_color is, so if user set only the latter,
an error occurred.

add_value() assumed that border_width is always < width
which led to awesome hanging up in an infinite loop,
should that fail to be the case.

The graph border_width property was driven
by the border_color property, instead of the other way round.

If border_color wasn't set, the border wasn't drawn all,
even if user sets non-zero border_width, which is
a confusing behavior, in contrast to all other widgets,
which simply use a plethora of fallback colors
to draw the border, when needed.

This also meant that, despite being documented,
the beautiful.graph_border_color fallback color
could not and was not ever used.

If the border_color was set though, an unset
border_width was immediately assumed to be 1 instead of 0,
also contrary to what other widgets do.

Theme colors shouldn't influence widget layout,
so this commit fixes this behavior to:

1) border_width always defaults to 0, if unset.
2) border_color falls back to
   beautiful.graph_border_color or white.

Maybe one should also add a beautiful.graph_border_width
theme variable, like some widgets do.