Calculate max values num in graph:add_value() more precisely
At most so many values were kept in the widget values array,
as there are pixels of usable widget width (i.e. sans border).

This is exactly right for the surely most common case
of graphs with 1px-wide bars, but is more than needed for
graphs with step_width+step_spacing>1, and not enough for
graphs with subpixel bars.

Moreover this made graph autoscaling behave weirdly in
graphs with thicker bars, because
old values survived longer in the array and
kept affecting min_value/max_value calculations
even after they were shifted off-screen.

The maximum number of kept values is now calculated exactly
as the number of values that can be displayed at once,
taking step_width and step_spacing into account.

It might be worth adding a max_values_num property to the graph
to let user override this calculation, in case they draw
shapes in step_shape() outside of its supposed rectangle,
so our assumptions of visibility are wrong, or
the widget often gets resized, resulting in the annoyance of
visibly truncated series.