Simplify internal values representation in graph widget
graph._private.values was either a table of values or
a table of tables of values, depending on whether
"stack" property is enabled, ... or a mix of both
if user forgot to pass the second "group" parameter
to graph:add_value().

This resulted in pretty fragile and confusing behavior,
e.g. leaving user to wonder, why they see nothing more
after setting stack=true for a graph in a previously working
rc.lua, and demanding a very careful use of
add_value() from the user, or else some values or
even entire stacks would mysteriously vanish
and errors in draw() get triggered,
sometimes after a period of seemingly normal functioning.

Now `values` is always a table of tables and
the group parameter in add_value() defaults to 1,
thus making the non-stacked graph a trivial variant
of the stacked graph.