Simplify internal values representation in graph even more
New values in graph._private.values were pushed to the end,
and old values were removed from the beginning of the table.

In draw() however newer values are drawn first,
at the left edge of the graph, necessitating
annoying idx = #values - idx conversions everywhere.

In particular in stacked graph that means that
values that are drawn together may have different indices.

One has to shift the values array in add_value()
either way, so one might as well insert
new values at the beginning of the array
and simplify the rest of the code.

This doesn't seem like much yet, but it will prove
to be much more convenient and less error-prone in
what I'm going to do later.