Use rectangles for drawing graph bars of all sizes
I have benchmarked various bar drawing strategies
and graph sizes and found out that drawing graph bars
with rectangles is as fast or even marginally faster
than using fat lines.

Thus the code that tries to special-case 1px bars
as lines seems to be not worth the additional complexity
it introduces, and this commit removes it.

Incidentally, I've also established, that rectangles
are more robust. If one draws graph bars and feeds
cairo very large coordinates (~2^62), then corresponding
rectangles fail to render, but all the other rectangles
in the same series get rendered successfully,
whereas a single large coordinate in a move_to/line_to()
sequence is enough to cause all lines in the batch
to fail to render.