Move value clamping from graph:add_value() into draw()
Previously all values were forcibly clamped between
0 (not tunable at all) and max_value (1 by default).

Thus graphs with negative values were not an option.
NaN values were conflated with zeros, even though
it may be useful to maintain the distinction and render
them differently. And if user happened to dynamically
change the max_value property, they discovered,
that their data got truncated.

Preserving user data as it was given is more flexible and
makes things easier to debug. This commit does exactly that.

So what happens now to the newly allowed values during draw()?
There are 3 new cases to consider:
1) NaNs
2) Negatives
3) Values outside of min_value..max_value range

NaNs are presently simply not drawn. An empty place is reserved
where the value should've been, but nothing else is done.
This is distinct from how a zero is handled,
e.g. if step_shape is set, it is called with the height
parameter = 0, which can still end up drawing something,
depending on the shape.
One could also consider addition of nan_color/nan_shape
properties which could be drawn to alert user
to a gap in their data.

Negative values necessitate introduction of the concept of
the zero axis, from which they will be drawn DOWN, in contrast to
positive values. Where should this axis be drawn?
How should step_shape-s with inherent direction, e.g. arrows,
be drawn for negatives? This will be addressed in the next commit.

Values outside of min/max_value range can imply shapes that
go beyond widget dimensions. An arrow shape that goes beyond
the roof is visibly different from an arrow that saturates and
merely touches the widget top, and the user may reasonably desire
either behavior. Therefore this commit introduces
a boolean clamp_bars property, which defaults to true
for reasons of robustness and backward compatibility with
the behaviour prior to this commit. The property of course affects
only the display of the values, the actual user data stays intact.