Add baseline_y and baseline_value properties to graph widget
Now values in graph have a *baseline*, tunable with
baseline_y and baseline_value properties.

It is the (for now invisible) horizontal line from which
all bars in the graph grow. I call it a "baseline"
instead of "axis", because it doesn't actually have to correspond
to the 0 value. But it can, and actually does now by default.

baseline_y and baseline_value properties determine
the vertical placement of the baseline in the widget.

baseline_y is the position of the baseline in normalized
widget coordinates, e.g. 0 corresponds to widget's top,
1 - to widget's bottom, and 0.5 - to widget's middle.
So if e.g. one sets baseline_y = 0, then no matter how the widget
is resized and how its values are scaled, bars in it will be drawn
starting from the top.

baseline_value is the position of the baseline in
value coordinates.
So if one sets baseline_value = 10, then no matter what happens,
bars for values greater than 10 will be drawn as growing up from
this line, and bars for values smaller as growing down.
Where and if it will be seen in the widget, depends on the scaling.
baseline_value = 0 is the familiar zero axis.

baseline_y is unset by default and has priority, i.e.
if it's set, then baseline_value is ignored.
baseline_value defaults to 0.

Prior to this commit, bars were always drawn from the bottom edge
of the widget, or, in the new terms, the widget had baseline_y = 1.
But it was indistinguishable from baseline_value = 0, because
the widget clamped all negative values to 0, so the difference
between the two could never be observed. baseline_value = 0 is
clearly more useful and familiar to people, so now it is in force
by default.