Extract values size computation into compute_drawn_values_num()
This is the first step of making add_values() not
concern itself (directly) with widget dimensions.

The user can monkey-patch the method to override
how many values are (logically) drawn, because
it affects graph auto-scaling and can be useful for rendering
of step_shape-s that paint outside of their step_width.

I didn't make it a full-blown function-valued property,
because it seems a fringe functionality,
which I'm too lazy to document.