Track graph draw() usage to guess necessary array capacity
The last step of divorcing add_values() from width calculations,
while ensuring that user doesn't observe a lack of data to draw.

The widget now tracks an approximate number of values that
it has been asked to draw() in _private.last_drawn_values_num,
and truncates its values array based on that
(if not overridden by the capacity property).