Refactor graph:draw() to separate concerns
Firstly, summation of group values for stacked graphs
is extracted into graph:preprocess_values().

This makes graph:draw() smaller and exposes the fact, that
stacked graphs are almost entirely just a form of
a more general concept: pre-render data manipulation.
It's something a user might want to customize,
and now they can do it by monkey-patching the method.

Secondly, the logic for deciding, if and with which color
a data group should be drawn, is extracted into
graph:should_draw_data_group() and get_data_group_color()
methods respectively.

This makes it customizable too and makes more obvious
the fact, that it is a general thing that has nothing
to do with stacked graphs in particular.

With this there are only 5 lines remaining in graph:draw(),
that could be said to be specific to stacked graphs,
all pertaining to the `prev_y` variable, which merely
preserves the bar coordinates from a previous data group.