Implement graph.nan_color and nan_indication properties
Now, whenever there's a NaN among values and
graph.nan_indication is set, the corresponding area,
where the value bar should have been drawn, is filled
with the nan_color from top to bottom.

The painting is done after all data is rendered,
to make sure that it won't be overpainted and go unnoticed.

NaNs among graph's values are inevitably a sign of
some error and it is arguably not a sane default behavior
to simply not render them at all. The user should take
immediate note of any errors by default, instead of
wondering at a mysteriously empty graph widget, which
is why nan_indication is set to true by default.
If the user wants to silence NaNs, they can always simply
set it to false.

nan_color if unset, defaults to a noticeable yellow-black
pattern.

As a direct and planned consequence, stacked graphs
with negative values in them trigger the NaN handling too.
But silencing NaN handling for those graphs is also possible
and semi-sane, because then NaNs and negatives are ignored
and the graph behaves (almost) like they were zeros.